+++ RELEASE_NOTES.md
@@ -5,6 +5,10 @@ Changes:
   * Improved performance of bulk QC pages
   * Removed bulk Create/Edit QC Types pages - QC Types can now only be created and modified using the
     single Create/Edit QC Type page
+  * Changed the following library fields to be required for propagated (not received) libraries:
+    * Kit lot
+    * Thermal cycler - if there are thermal cyclers configured
+    * SOP - if there are SOPs configured
   * Advanced search now works in tables on Edit Workset page
   * Added back missing "Check QCs" button on bulk library pages
   * Fixed receiving libraries from the Edit Project page